+++ ricaun.RevitTest.Tests/TestsAsync.cs
@@ -1,12 +1,33 @@
 ﻿using Autodesk.Revit.UI;
 using NUnit.Framework;
+using ricaun.Revit.Async;
 using System;
 using System.Threading.Tasks;
 
 namespace ricaun.RevitTest.Tests
 {
+    public class TestRevitTask
+    {
+        [Test]
+        public void Initialize()
+        {
+            RevitTask.Initialize();
+        }
+    }
+
     public class TestsAsync
     {
+        [Test]
+        public async Task TestAsync_MacroManager()
+        {
+            await RevitTask.Run((uiapp) =>
+            {
+                uiapp.PostCommand(RevitCommandId.LookupPostableCommandId(PostableCommand.MacroManager));
+            });
+            var inContext = await RevitTask.Run((uiapp) => { return InContext(uiapp); });
+            Assert.IsTrue(inContext);
+        }
+
         [Test]
         public async Task TestAsync()
         {
@@ -18,7 +39,12 @@ namespace ricaun.RevitTest.Tests
         public async Task TestAsync(UIApplication uiapp)
         {
             await Task.Delay(100);
-            Console.WriteLine(InContext(uiapp));
+            var inContextShouldBeFalse = InContext(uiapp);
+            Console.WriteLine(inContextShouldBeFalse);
+            Assert.IsFalse(inContextShouldBeFalse);
+            var inContext = await RevitTask.Run((app) => { return InContext(app); });
+            Console.WriteLine(inContext);
+            Assert.IsTrue(inContext);
         }
 
         private bool InContext(UIApplication uiapp)